# offchain/backend-go/docs/swagger.yaml
basePath: /v1
definitions:
...
  github_com_vndc_backend_internal_models.BalanceData:
    properties:
      available:
        description: Available is OnChain minus Pending — the amount the user can
          spend right now.
        example: "4000000000000000000000"
        type: string
      on_chain:
        description: OnChain is the confirmed on-chain balance in wei (as decimal
          string).
        example: "5000000000000000000000"
        type: string
      pending:
        description: Pending is the sum of outgoing wei reserved for queued transfers
          not yet settled.
        example: "1000000000000000000000"
        type: string
      synced_at:
        description: SyncedAt is when the on-chain balance was last fetched (RFC3339).
        example: "2026-04-24T07:00:00Z"
        type: string
      wallet:
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        type: string
    type: object
  github_com_vndc_backend_internal_models.BalanceResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.BalanceData'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_vndc_backend_internal_models.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details: {}
      message:
        example: wallet query param required
        type: string
    type: object
  github_com_vndc_backend_internal_models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: false
        type: boolean
    type: object
...
  github_com_vndc_backend_internal_models.PaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      pages:
        example: 5
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  github_com_vndc_backend_internal_models.ResponseMeta:
    properties:
      request_id:
        example: 01HWXYZ1234567890ABCDEF
        type: string
      timestamp:
        example: "2026-04-24T07:00:00Z"
        type: string
    type: object
...
  github_com_vndc_backend_internal_models.TransactionInfo:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      batch_id:
        example: batch-001
        type: string
      block_number:
        example: 19500000
        type: integer
      created_at:
        example: "2026-04-24T07:00:00Z"
        type: string
      deadline:
        example: 1745481600
        type: integer
      from_wallet:
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      nonce:
        example: "42"
        type: string
      retry_count:
        example: 0
        type: integer
      settled_at:
        type: string
      signature:
        example: 0xabc123...
        type: string
      status:
        example: PENDING
        type: string
      to_wallet:
        example: 0xAbCdEf1234567890AbCdEf1234567890AbCdEf12
        type: string
      tx_hash:
        example: 0xdeadbeef...
        type: string
      type:
        example: TOKEN_TRANSFER
        type: string
      updated_at:
        example: "2026-04-24T07:05:00Z"
        type: string
    type: object
  github_com_vndc_backend_internal_models.TransactionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.TransactionInfo'
        type: array
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      pagination:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.PaginationMeta'
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.TransactionResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.TransactionInfo'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.TransferRequest:
    properties:
      amount:
        description: Amount is the transfer value in wei expressed as a decimal string
          (no floats).
        example: "1000000000000000000"
        type: string
      deadline:
        description: Deadline is a Unix timestamp after which the signed transfer
          is invalid.
        example: 1745481600
        type: integer
      from_wallet:
        description: FromWallet is the sender's checksummed Ethereum address (must
          match JWT).
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        type: string
      nonce:
        description: Nonce is the EIP-712 nonce (decimal string). Must be unique per
          sender.
        example: "42"
        type: string
      signature:
        description: Signature is the 0x-prefixed 65-byte EIP-712 signature (132 hex
          chars + 0x prefix).
        example: 0x4a5b6c...130hexchars...ef01
        type: string
      to_wallet:
        description: ToWallet is the recipient's checksummed Ethereum address.
        example: 0xAbCdEf1234567890AbCdEf1234567890AbCdEf12
        type: string
    type: object
  github_com_vndc_backend_internal_models.TransferResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.TransactionInfo'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@vndc.io
    name: VNDC Platform Team
  description: |-
    Off-chain backend for the VNDC token platform.

    ## Authentication
    The API uses **wallet-based authentication** (Sign-In With Ethereum / EIP-191).
    All protected endpoints require an `Authorization: Bearer <access_token>` header.

    ### Login flow
    1. `GET /v1/auth/challenge?wallet=0x…`  — obtain a nonce-embedded message
    2. Sign the message with your wallet (personal_sign / EIP-191)
    3. `POST /v1/auth/login` — submit the signature; receive a JWT pair
    4. *(if 2FA is enabled)* `POST /v1/auth/2fa/complete` — verify TOTP code
    5. Pass the `access_token` as `Authorization: Bearer <token>` on every request
    6. `POST /v1/auth/refresh` — rotate tokens before access token expires (15 min)

    ## Amount encoding
    All token amounts are expressed in **wei** as **decimal strings** — never floats.
    Example: `"1000000000000000000"` = 1 VNDC (18 decimals).
  license:
    name: Proprietary
    url: https://vndc.io/terms
  title: VNDC Backend API
  version: 1.0.0
paths:
...
  /tokens/balance/{wallet}:
    get:
      description: |-
        Returns the on-chain confirmed balance, pending (unprocessed transfers) amount,
        and spendable (available) balance for a given Ethereum wallet address.

        This endpoint is **public** — no authentication required.
      parameters:
      - description: Checksummed Ethereum wallet address
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        in: path
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.BalanceResponse'
        "400":
          description: Invalid wallet address format
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: Wallet not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      summary: Get wallet token balance
      tags:
      - Tokens
  /tokens/transactions:
    get:
      description: |-
        Returns a paginated list of token transfers where the authenticated user is
        either the sender or the recipient. Sorted by creation time, newest first.
      parameters:
      - description: 'Page number (default: 1)'
        example: 1
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        example: 20
        in: query
        name: page_size
        type: integer
      - description: Filter by status
        enum:
        - PENDING
        - COMPLETED
        - FAILED
        - CANCELLED
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.TransactionListResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my transactions
      tags:
      - Tokens
  /tokens/transactions/{id}:
    get:
      description: |-
        Returns the full details for a single transfer. Users may only view transactions
        where they are the sender or recipient. Admins may view any transaction.
      parameters:
      - description: Transaction ID (MongoDB ObjectID or UUID)
        example: 65f1a2b3c4d5e6f7a8b9c0d2
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.TransactionResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Access denied — not sender or recipient
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: Transaction not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a transaction by ID
      tags:
      - Tokens
  /tokens/transfer:
    post:
      consumes:
      - application/json
      description: |-
        Processes an **EIP-712** (or EIP-191 personal_sign) off-chain token transfer.
        The signature is verified server-side using the sender's wallet address.

        **Replay protection**: Each `(from_wallet, nonce)` pair must be globally unique.
        Use a monotonically increasing nonce or a random UUID converted to uint256.

        **Deadline**: Unix timestamp (seconds). Transactions submitted after `deadline` are rejected.
      parameters:
      - description: EIP-712 signed transfer payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.TransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transfer queued / recorded
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.TransferResponse'
        "400":
          description: Missing fields / past deadline
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Invalid EIP-712 signature
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "402":
          description: Insufficient balance
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "409":
          description: Duplicate nonce
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a signed token transfer
      tags:
      - Tokens
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...